refactor: reorg datamod
