+++ concrete/codegen/datamod/datamod.go
@@ -26,88 +26,10 @@ import (
 	"strings"
 	"text/template"
 	"unicode"
-
-	"github.com/ethereum/go-ethereum/concrete/lib"
 )
 
 //go:embed struct.tpl
 var structTpl string
-
-type StorageStruct struct {
-	storage lib.SlotArray
-	offsets []int
-	sizes   []int
-}
-
-func NewStorageStruct(slot lib.StorageSlot, sizes []int) *StorageStruct {
-	offsets := make([]int, len(sizes))
-	offset := 0
-	nSlots := 0
-	for i := 1; i < len(sizes); i++ {
-		size := sizes[i]
-		if offset/32 < (offset+size)/32 {
-			offset = (offset/32 + 1) * 32
-			nSlots++
-		}
-		offset += size
-		offsets[i] = offset
-	}
-	return &StorageStruct{
-		storage: slot.SlotArray([]int{nSlots}),
-		offsets: offsets,
-		sizes:   sizes,
-	}
-}
-
-func (s *StorageStruct) GetField(index int) []byte {
-	absOffset := s.offsets[index]
-	slotIdx := absOffset / 32
-	slotOffset := absOffset % 32
-
-	slotValue := s.storage.Value(slotIdx).Bytes32()
-	size := s.sizes[index]
-	return slotValue[slotOffset : slotOffset+size]
-}
-
-func (s *StorageStruct) SetField(index int, data []byte) {
-	absOffset := s.offsets[index]
-	slotIdx := absOffset / 32
-	slotOffset := absOffset % 32
-
-	slot := s.storage.Value(slotIdx)
-	slotValue := slot.Bytes32()
-	size := s.sizes[index]
-	copy(slotValue[slotOffset:slotOffset+size], data)
-	slot.SetBytes32(slotValue)
-}
-
-type FieldSchema struct {
-	Name  string
-	Title string
-	Index int
-	Type  FieldType
-}
-
-type MappingSchema struct {
-	Name   string
-	Keys   []FieldSchema
-	Values []FieldSchema
-}
-
-type ModelSchema []MappingSchema
-
-type MappingUnmarshal struct {
-	KeySchema map[string]string `json:"keySchema"`
-	Schema    map[string]string `json:"schema"`
-}
-
-type ModelUnmarshal map[string]MappingUnmarshal
-
-type Config struct {
-	JSON    string
-	Out     string
-	Package string
-}
 
 func lowerFirstLetter(str string) string {
 	if len(str) == 0 {
@@ -135,6 +57,28 @@ func isValidName(name string) bool {
 	return re.MatchString(name) && len(strings.TrimSpace(name)) == len(name)
 }
 
+type FieldSchema struct {
+	Name  string
+	Title string
+	Index int
+	Type  FieldType
+}
+
+type MappingSchema struct {
+	Name   string
+	Keys   []FieldSchema
+	Values []FieldSchema
+}
+
+type ModelSchema []MappingSchema
+
+type MappingUnmarshal struct {
+	KeySchema map[string]string `json:"keySchema"`
+	Schema    map[string]string `json:"schema"`
+}
+
+type ModelUnmarshal map[string]MappingUnmarshal
+
 func newFieldSchema(name string, index int, typeStr string) (FieldSchema, error) {
 	if !isValidName(name) {
 		return FieldSchema{}, fmt.Errorf("invalid name: %s", name)
@@ -151,6 +95,44 @@ func newFieldSchema(name string, index int, typeStr string) (FieldSchema, error)
 	}, nil
 }
 
+func unmarshalModel(jsonContent []byte) (ModelSchema, error) {
+	var unmarshaledModel ModelUnmarshal
+	err := json.Unmarshal(jsonContent, &unmarshaledModel)
+	if err != nil {
+		return ModelSchema{}, err
+	}
+
+	var model ModelSchema
+	for name, mapping := range unmarshaledModel {
+		if !isValidName(name) {
+			return ModelSchema{}, fmt.Errorf("invalid name: %s", name)
+		}
+		newMapping := MappingSchema{Name: upperFirstLetter(name)}
+		for keyName, keyType := range mapping.KeySchema {
+			fieldSchema, err := newFieldSchema(keyName, len(newMapping.Keys), keyType)
+			if err != nil {
+				return ModelSchema{}, err
+			}
+			newMapping.Keys = append(newMapping.Keys, fieldSchema)
+		}
+		for valueName, valueType := range mapping.Schema {
+			fieldSchema, err := newFieldSchema(valueName, len(newMapping.Values), valueType)
+			if err != nil {
+				return ModelSchema{}, err
+			}
+			newMapping.Values = append(newMapping.Values, fieldSchema)
+		}
+		model = append(model, newMapping)
+	}
+	return model, nil
+}
+
+type Config struct {
+	JSON    string
+	Out     string
+	Package string
+}
+
 func GenerateDataModel(config Config) error {
 	if !isValidName(config.Package) {
 		return fmt.Errorf("invalid package name: %s", config.Package)
@@ -160,40 +142,14 @@ func GenerateDataModel(config Config) error {
 	if err != nil {
 		return err
 	}
-
-	var unmarshaledModel ModelUnmarshal
-	err = json.Unmarshal(jsonContent, &unmarshaledModel)
+	model, err := unmarshalModel(jsonContent)
 	if err != nil {
 		return err
-	}
-
-	var model ModelSchema
-	for name, mapping := range unmarshaledModel {
-		if !isValidName(name) {
-			return fmt.Errorf("invalid name: %s", name)
-		}
-		newMapping := MappingSchema{Name: upperFirstLetter(name)}
-		for keyName, keyType := range mapping.KeySchema {
-			fieldSchema, err := newFieldSchema(keyName, len(newMapping.Keys), keyType)
-			if err != nil {
-				return err
-			}
-			newMapping.Keys = append(newMapping.Keys, fieldSchema)
-		}
-		for valueName, valueType := range mapping.Schema {
-			fieldSchema, err := newFieldSchema(valueName, len(newMapping.Values), valueType)
-			if err != nil {
-				return err
-			}
-			newMapping.Values = append(newMapping.Values, fieldSchema)
-		}
-		model = append(model, newMapping)
 	}
 
 	funcMap := template.FuncMap{
 		"sub": func(a, b int) int { return a - b },
 	}
-
 	tmpl, err := template.New("struct").Funcs(funcMap).Parse(structTpl)
 	if err != nil {
 		return err
